@@ -17,15 +17,20 @@
     "d3-selection": "^1.3.2",
     "d3-shape": "^1.2.2",
     "d3-transition": "^1.1.3",
+    "i18next": "^17.0.3",
     "lodash": "^4.17.11",
-    "react": "^16.4.2",
-    "react-dom": "^16.4.2",
+    "react": "^16.11.0",
+    "react-dom": "^16.11.0",
+    "react-i18next": "^10.11.0",
     "react-redux": "^6.0.1",
-    "react-router-dom": "^4.3.1",
-    "react-scripts": "^2.0.5",
+    "react-router-dom": "^5.1.2",
+    "react-scripts": "^3.0.1",
     "react-slidedown": "^2.0.9",
+    "video-react": "^0.14.1",
     "redux": "^4.0.1",
+    "redux-thunk": "^2.3.0",
     "seedrandom": "^3.0.1",
+    "socket.io-client": "^2.2.0",
     "tinycolor2": "^1.4.1"
   },
   "scripts": {
