@@ -1,37 +1,42 @@
-import React, { Component } from 'react'
+import React, { useState, useRef, useEffect } from 'react'
 import _ from 'lodash'
 import { select } from 'd3-selection'
+import { scaleBand } from 'd3-scale'
 import AnimatedText from './AnimatedText'
 import { animationTime } from './Utils'
+import COLORS from '../colors'
 
-class Bar extends Component {
-	constructor(props) {
-        super(props)
-		this.node = React.createRef()
-		this.state = { x: props.x, y: props.y, width: props.width, height: props.height }
-	}
+export const Bar = ({ fadeIn = true, ...props }) => {
+	const node = useRef(),
+	[x] = useState(props.x), 
+	[y] = useState(props.y), 
+	[width] = useState(props.width),
+	[height] = useState(props.height),
+	[fill] = useState(props.fill)
 
-	componentDidUpdate() {
-		select(this.node.current)
-			.datum(this.props)
+	useEffect(() => {
+		select(node.current)
+			.datum(props)
             .transition()
 			.duration(animationTime)
 			.attr('x', d => d.x)
 			.attr('y', d => d.y)
 			.attr('height', d => d.height)
 			.attr('width', d => d.width)
-	}
+			.attr('fill', d => d.fill)
+	})
 
-	render() {
-		return (<rect
-			ref={this.node}
-			style={this.props.style}
-			x={this.state.x}
-			y={this.state.y}
-			height={this.state.height}
-			width={this.state.width} 
-			onClick={this.props.onClick} />)
-	}
+	return <rect
+		ref={node}
+		fill={fill}
+		style={props.style}
+		x={x}
+		y={fadeIn ? y + height : y}
+		height={fadeIn ? 0 : height}
+		width={width} 
+		onClick={props.onClick}
+		opacity={.9}
+	/>
 }
 
 const BarGraph = props => {
@@ -43,29 +48,34 @@ const BarGraph = props => {
 	const clickHandler = props.onClick || (() => {})
 	return (
 	<g className={`bar-graph ${props.className || ''}`}>
-	{props.data.map((d, i) => 
-		[].concat(d.y).map((y,j) => 
-		<g key={"bar" + d.x + j} className="bar animated" style={props.style}>
-			<Bar
-				style={{fill: d.highlight ? (props.highlightColor || '#fe99f2') : (_.isArray(d.color) ? d.color[j] : d.color || props.color || '#fe9922')}} 
-				x={props.xScale(d.x) + dx}
-				y={props.yScale(y)}
-				height={props.yScale.range()[0] - props.yScale(y)}
-				width={width} 
-				onClick={() => clickHandler(d, i)} />
-			)}
-			{props.labels && 
-				<AnimatedText 
-					x={props.xScale(d.x) + (props.xScale.bandwidth ? (width/2) : offset)} 
+	{props.data.map((d, i) => {
+		const values = [].concat(d.y)
+		const scale = scaleBand()
+			.domain([0, values.length-1])
+			.range([props.xScale(d.x) + dx, props.xScale(d.x) + dx + width])
+			.paddingOuter(0)
+		return <g key={d.x} className="bar animated" style={props.style}>
+			{values.map((y,j) => 
+			<g key={j}>
+				<Bar
+					fill={ d.highlight ? (props.highlightColor || COLORS.default) : (_.isArray(d.color) ? d.color[j] : d.color || props.color || COLORS.default)}
+					x={scale(j)}
+					y={props.yScale(y)}
+					height={props.yScale.range()[0] - props.yScale(y)}
+					width={scale.bandwidth()} 
+					onClick={() => clickHandler(d, i)} />
+				{props.labels && 
+					<AnimatedText 
+					startPos={{y:props.yScale(y) - 3 + props.yScale.range()[0] - props.yScale(y)}}
+					x={scale(j) + scale.bandwidth()/2 } 
 					y={props.yScale(y) - 3}>
-					{_.isArray(d.label) ? d.label[j] : d.label || y}
-				</AnimatedText>
-			}
+						{_.isArray(d.label) ? d.label[j] : d.label || y}
+					</AnimatedText>
+				}
+			</g>)}
 		</g>
-		)
-	)}
+	})}
 	</g>)
 }
 
 export default BarGraph
-export { Bar }
