@@ -1,42 +1,45 @@
-import React, { Component } from 'react'
+import React, { useState, useRef } from 'react'
+import _ from 'lodash/fp'
 import { OrdinalChart } from '../Charting/Chart'
 import { XAxis, YAxis } from '../Charting/Axis'
 import BarGraph from '../Charting/BarGraph'
+import { withTranslation } from 'react-i18next'
+import ColorLegend from '../Charting/ColorLegend'
+import COLORS from "../colors";
 const exams = ['alle MCs', 'letzter PTM']
 
-class Subject extends Component {
-    constructor(props) {
-        super(props)
-        this.state = { mode: "current", flash: props.flash }
-        this.node = React.createRef()
-    }
-    
-    componentDidMount() {
-        if(this.state.flash) window.scrollTo(0, this.node.current.getBoundingClientRect().y + 70)
-        this.setState({ flash: false })
+const colors = [COLORS.graphs.total, COLORS.progress.darker1]
+const Subject = ({t, ...props}) => {
+    const [ flash, setFlash ] = useState(props.flash)
+    const node = useRef(null)
+
+    if(flash) {
+        setTimeout(() => { 
+            document.querySelector('div.App').scrollTo(0, node.current.getBoundingClientRect().y - 70)
+            setFlash(false)
+        }, 500)
     }
 
-    setMode(mode) {
-        this.setState({ mode })
-    }
-
-    render() {
-        const data = exams.map((d, i) => ({ x:d, y: [this.props.data[i].correct, this.props.data[i].questions]}))
-        return (
-            <div className="card m-2 flex-grow-1 with-shadow" style={{width: '20rem'}}>
-                <div ref={this.node} className={`card-body ${this.state.flash ? 'bg-primary' : ''}`} style={{transition: '5s'}}>
-                    <span className="font-weight-bold">{this.props.title}</span>
-                    <div className="p-4">
-                        <OrdinalChart style={{height:'15rem'}} xDomain={exams} yDomain={[0,30]}>
-                            <YAxis label="gestellte vs. richtige Fragen" ticks={{count: 4}} />
-                            <XAxis />
-                            <BarGraph labels data={ data } color="hsla(250, 100%, 50%, .6)" />
-                        </OrdinalChart>
-                    </div>
+    const data = exams.map((d, i) => ({ color: colors, x:d, y: [props.data[i].maximalPunktzahl || 0, props.data[i].value || 0]}))
+    const max = Math.max(..._.map(d => _.ceil(d.maximalPunktzahl * 1.1) || 0, props.data))
+    return (
+        <div className="card m-2 flex-grow-1" style={{minWidth: '20rem', maxWidth: '40rem'}}>
+            <div ref={node} className={`card-body ${flash ? 'bg-primary' : ''}`} style={{transition: '5s'}}>
+                <span className="font-weight-bold">{props.titel}</span>
+                <div className="p-4">
+                <div className='mb-1'>
+                    <ColorLegend text={t('gestellte Fragen')} style={{backgroundColor: COLORS.graphs.total }}/>
+                    <ColorLegend text={t('richtige Antworten')} style={{backgroundColor: COLORS.progress.darker1 }}/>
+                </div>
+                    <OrdinalChart style={{height:'15rem'}} xDomain={exams} yDomain={[0,max]}>
+                        <YAxis ticks={{count: Math.min(max, 4)}} />
+                        <XAxis />
+                        <BarGraph labels data={ data } />
+                    </OrdinalChart>
                 </div>
             </div>
-        )
-    }
+        </div>
+    )
 }
 
-export default Subject
+export default withTranslation()(Subject)
