@@ -1,67 +1,49 @@
-import React, { Component } from 'react'
-import { BrowserRouter } from 'react-router-dom'
-import { Route, Redirect } from 'react-router'
-import { Provider } from 'react-redux'
-import { createStore } from 'redux'
+import React from 'react'
+import {BrowserRouter, NavLink} from 'react-router-dom'
+import {Route, Redirect} from 'react-router'
+import {Provider} from 'react-redux'
+import {createStore, compose, applyMiddleware} from 'redux'
+import thunkMiddleware from 'redux-thunk'
 import rootReducer from './reducer'
 import 'd3-transition'
 import 'react-slidedown/lib/slidedown.css'
+import 'video-react/dist/video-react.css'
 import './App.css'
-import Navbar from './Core/navbar'
-import Login from './Login'
-import Breadcrumbs from './Core/Breadcrumbs'
-import PrivateRoute from './Core/PrivateRoute'
-import Routes from './Core/Routes'
+import Feedback from './Core/Feedback'
+import Navbar from './Core/routing/navbar'
+import Breadcrumbs from './Core/routing/Breadcrumbs'
+import PrivateRoute from './Core/routing/PrivateRoute'
+import Routes from './Core/routing/Routes'
+import {withTranslation} from 'react-i18next'
 
-const getBasename = () => "/" + (window.location.pathname.split( '/' )[1] || "")
+const basename = process.env.PUBLIC_URL || '/'
 
-class App extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {loggedIn: true};
-  }
+const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose
 
-  render() {
-    let navbar = (
-      <Navbar isLoggedIn={this.state.loggedIn}></Navbar>
-    )
-    let breadcrumbs =  (
-<Breadcrumbs />
-    )
-
-    let login
-    if(!this.state.loggedIn) {
-      login = () => (<Login handleLogin={() => this.setState({loggedIn: true})}></Login>);
-      navbar = false
-      breadcrumbs = false
-    } else {
-      login = () => (<Redirect to="/dashboard"></Redirect>);
-    }
-
-    return (
-      <Provider store={createStore(rootReducer, window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__())}>
-        <BrowserRouter basename={getBasename()}>
-          <div className="App">
-
-            {navbar}
-            {breadcrumbs}
-
-            <Route path="/login" component={login} />
-            {Routes.map( route => ( route.private ?
-                    <PrivateRoute key={route.path} path={route.path} component={route.component} exact={route.exact} isLoggedIn={this.state.loggedIn} /> :
-                    <Route key={route.path} path={route.path} component={route.component} exact={route.exact} />
-            ))}
-            <Route exact path="/" render={() => (
-                <Redirect to="/dashboard" />
-            )}/>
-          </div>
+const App = withTranslation()(() =>
+    <Provider store={createStore(rootReducer, composeEnhancers(applyMiddleware(thunkMiddleware)))}>
+        <BrowserRouter basename={basename}>
+            <div className="App container-fluid p-0 d-flex flex-column">
+                <Navbar />
+                <Breadcrumbs/>
+                <div className="container-fluid flex-fill p-0">
+                    {Routes.map(route => (route.private ?
+                            <PrivateRoute key={route.path} path={route.path} component={route.component}
+                                            exact={route.exact} /> :
+                            <Route key={route.path} path={route.path} component={route.component}
+                                    exact={route.exact}/>
+                    ))}
+                    <Route exact path="/" render={() => (
+                        <Redirect to="/dashboard"/>
+                    )}/>
+                </div>
+                <div className="w-100 text-center" style={{fontSize: '.9rem', color: 'grey', backgroundColor: 'lightgrey'}}>
+                    <NavLink to="/impressum">Impressum / Disclaimer</NavLink>
+                </div>
+                <Feedback />
+            </div>
         </BrowserRouter>
-      </Provider>
-    );
-  }
+    </Provider>
+)
 
-
-
-}
-
-export default App;
+export default App
