@@ -1,12 +1,31 @@
 import React from 'react'
+import {FontAwesomeIcon} from '@fortawesome/react-fontawesome'
+import {faCheck} from '@fortawesome/free-solid-svg-icons'
 
-export default function Checkbox(props) {
-    return (<div className="input-group col-sm-4">
-            <div className="input-group-prepend" onClick={() => props.onChange()}>
-                <div className="input-group-text" style={{backgroundColor: `hsla(${props.color}, 100%, 60%, .4)`}}>
-                    <input type="checkbox" checked={props.checked} onChange={() => props.onChange()} />
-                </div>
-                <span className="input-group-text">{props.label}</span>
-            </div>
-        </div>)
+const Checkbox = ({ checked, lineColor, backgroundColor, children, onUncheck, onCheck }) => {
+    const toggle = () => {
+        if (checked)
+            onUncheck()
+        else
+            onCheck()
+    }
+
+    return (
+        <span onClick={toggle}>
+                <label className="p-2 m-0">
+                    <span className="mr-2 pl-1"
+                          style={{
+                              backgroundColor: backgroundColor,
+                              color: checked
+                                  ? lineColor
+                                  : backgroundColor
+                          }}>
+                         <FontAwesomeIcon className="mr-1" icon={faCheck}/>
+                    </span>
+                    {children}
+                </label>
+            </span>
+    )
 }
+
+export default Checkbox
