@@ -1,14 +1,16 @@
 import React from 'react'
 import Station from './Station'
+import COLORS from "../../colors";
 
-const borderColor = 'rgba(0,0,0,.35)'
 const StationsExam = props => (
-    <div className="mb-1 p-1" style={{border: `1px dashed ${borderColor}`}}>
-        <div className="text-right" style={{fontSize: '.75rem', color: borderColor}} >
+    <div className="mb-3 p-3 pt-2">
+        <div className="text-right" style={{fontSize: '.75rem'}} >
             {props.label}
         </div>
         {props.stations.map(station => 
-            <Station key={station.name} color={props.colors(station.category)} data={station} {...props.scales} />
+            <Station key={station.code} color={props.colors(station.zeitsemester)} data={station} {...props.scales}
+                colorTotal={COLORS.background.grey0} colorPartOfTotal={COLORS.background.grey6}
+            />
         )}
     </div>)
 
