@@ -1,48 +1,51 @@
-import React, { Component } from 'react'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { faCheckSquare, faSquare } from '@fortawesome/free-solid-svg-icons'
+import React from 'react'
+import Checkbox from "../Exams/Checkbox";
+import COLORS from "../colors";
 
-class Filter extends Component {
-    constructor( props) {
-        super(props)
-        this.state = { filters: props.filters }
+const Filter = ({ 
+    filters, 
+    colors = { background: COLORS.background.grey5,
+        text: COLORS.background.base },
+    ...props 
+}) => {
+
+    function callUpdate(filters) {
+        props.onUpdate(filters)
     }
 
-    toggleFilter(filter) {
-        filter.selected = !filter.selected
-        this.setState({ filters: this.state.filters }, this.callUpdate)
+    function toggleFilter(filter) {
+        const index = filters.indexOf(filter)
+        const newFilters = [ ...filters.slice(0, index), { ...filter, selected: !filter.selected }, ...filters.slice(index + 1) ]
+        callUpdate(newFilters)
     }
 
-    selectAll() {
-        this.setState({ filters: this.state.filters.map(filter => ({ ...filter, selected: true })) }, this.callUpdate)
+    function all(selected) {
+        const newFilters = filters.map( filter => ({ ...filter, selected }) )
+        callUpdate(newFilters)
     }
 
-    selectNone() {
-        this.setState({ filters: this.state.filters.map(filter => ({ ...filter, selected: false })) }, this.callUpdate)
-    }
-
-    callUpdate() {
-        this.props.onUpdate(this.state.filters)
-    }
-
-    render() {
-        return (
-            <div className="row mb-2 mt-1" style={this.props.style}>
-                <div className="col">
-                    {this.props.showAll && <button className="btn btn-outline-primary btn-sm mr-1 mt-1" onClick={ () => this.selectNone() }>alle anzeigen</button>}
-                    {this.state.filters.map(filter => (
-                        <span className="d-inline-block mr-1 mt-1" 
-                            key={filter.label} 
-                            style={{ fontSize: '.73rem', border: `1px solid ${filter.color || 'hsla(210, 50%, 55%, 1)'}`, backgroundColor: filter.color || 'hsla(210, 50%, 60%, 1)'}}>
-                            <label className="p-2 m-0 text-white" onClick={() => this.toggleFilter(filter)} >
-                                <FontAwesomeIcon className="mr-1" icon={filter.selected ? faCheckSquare : faSquare} />
-                            {filter.label}</label>
-                        </span>)
-                    )}
-                </div>
-            </div>
-        )
-    }
+    return (
+        <div className={`d-inline-flex flex-row flex-wrap ${props.className}`}>
+        {props.showAll && <button className="btn btn-outline-primary btn-sm mr-1 mt-1" onClick={ () => all(false) }>alle anzeigen</button>}
+        {filters.map(filter => (
+            <span className="flex-fill d-inline-block mr-1 mt-1" 
+                key={filter.label} 
+                style={{ 
+                    fontSize: '.73rem', 
+                    border: '1px solid ' + colors.line,
+                    backgroundColor: colors.background
+                }}>
+                <Checkbox checked={filter.selected}
+                          onCheck={() => toggleFilter(filter)}
+                          onUncheck={() => toggleFilter(filter)}
+                          lineColor={colors.background}
+                          backgroundColor={colors.text} >
+                          <span style={{color: colors.text}}>{filter.label}</span>
+                </Checkbox>
+            </span>)
+        )}
+        </div>
+    )
 }
 
 export default Filter
