@@ -1,141 +1,116 @@
-
-export const ConsultingTree = {
-    ExternConsulting : [
+export const ConsultingTree = t => ({
+    ExternConsulting: [
         {
-            title : 'Finanzierung des Studiums',
+            title: t('Finanzierung des Studiums'),
             href: 'https://www.stw.berlin/finanzierung.html'
         },
         {
-            title : 'Stipendien-möglichkeiten ',
-            href : 'https://www.stipendienlotse.de/datenbank.php?A%5B%5D=2&H=&B%5B%5D=4&F=57&S=&job=search'
+            title: t('Stipendien'),
+            href: 'https://www.stipendienlotse.de/datenbank.php?A%5B%5D=2&H=&B%5B%5D=4&F=57&S=&job=search'
         },
         {
-            title : 'Finanzielle Hilfen ',
-            href : 'https://www.stw.berlin/beratung/themen/finanzelle-hilfen.html'
-        },
-       
-        
-        {
-            title : 'Wohnungssuche',
-            href:'https://www.stw.berlin/wohnen.html#jump_wohnraumb%C3%B6rsen'
+            title: t('Finanzielle Hilfen '),
+            href: 'https://www.stw.berlin/beratung/themen/finanzelle-hilfen.html'
         },
         {
-            title : ' Studierenden-Jobs',
-            href:'https://www.stw.berlin/jobben.html#jump_jobs_&_trainings_f%C3%BCr_studierende'
+            title: t('Wohnungssuche'),
+            href: 'https://www.stw.berlin/wohnen.html#jump_wohnraumb%C3%B6rsen'
         },
         {
-            title : 'Hilfe bei Zweifel am Studium ',
-            href:'https://www.studienabbruch-und-dann.de/de/fragen-und-antworten-60.html'
+            title: t(' Studierenden-Jobs'),
+            href: 'https://www.stw.berlin/jobben.html#jump_jobs_&_trainings_f%C3%BCr_studierende'
         },
         {
-            title : 'Hilfe bei Studienabbruch ',
-            href:'https://www.queraufstieg-berlin.de'
+            title: t('Hilfe bei Zweifel am Studium '),
+            href: 'https://www.studienabbruch-und-dann.de/de/fragen-und-antworten-60.html'
         },
         {
-            title : 'Navigation Beratung (video) ',
-            href:'https://www.youtube.com/watch?time_continue=20&v=2nea4zUSkTY'
+            title: t('Hilfe bei Studienabbruch '),
+            href: 'https://www.queraufstieg-berlin.de'
         },
     ],
-    ConsultingCard : [
-    
-    {
-        title:"MediCoach",
-        paragraph:"MediCoach bietet  Ihnen ein individuelles, kostenloses, lösungsorientiertes Coaching zur Unterstützung stressbelasteter Studierender im Rahmen Ihres Studiums an der Charité Universitätsmedizin Berlin an. Neben Einzelcoachings können Sie auch in einer in jedem Semester angebotenen Coachinggruppe gemeinsam mit Ihren Kommiliton*innen Lösungen für Probleme im Studium entwickeln.",
-        name1:"Dipl.-Soz. Silke Boehm",
-        name2:"Dipl.-Soz. Sylvie Tappert",
-        tel:" +49 30 450 529189",
-        email1:"medicoach@charite.de",
-        address:"Luisenstr 57 CCM ",
-        talk:" Gespräch nach Terminvereinbarung",
-        href: 'https://medicoach.charite.de/fuer_studierende' 
-    }, 
-    {
-        title:"Familienbüro",
-        paragraph:"Sie studieren an der Charité und betreuen Kinder? Sie suchen einen Kitaplatz? Sie pflegen Angehörige? Sie haben Fragen und benötigen Informationen rund um das Thema Vereinbarkeit von Studium und Familie?",
-        name1:"Sabine Barleben(Leitung)",
-        name2:"Stefanie Schmidt(Sekretariat)",
-        tel:"+49 030 450 577 257",
-        email1:"familienbuero@charite.de",
-        email2:"Väterbeauftragte:  vaeterbeauftragte@charite.de",
-        address: "Westring 1, 1. Etage, R1.1011 CVK",
-        talk:"CVK: dienstags von 11:00 bis 13:00 Uhr, Gästehaus R.0.0120",          
-        talk1:"CCM: mittwochs von 11:00 bis 13:00 Uhr, Hufelandweg 2, EG, Raum 56|57",
-        href: 'https://familienbuero.charite.de/charite_mit_kind/studieren_mit_familie/kinderbetreuung' 
-    },
-   {
-        title:"ChiC",
-        paragraph:"ChiC bietet Informationen und Unterstützung rund um die Themen Studieren im Ausland bzw. aus dem Ausland und Unterstützung von internationalen Studierenden in Berlin.",
-        name1:"Lutz Steiner(Leitung)",
-        name2:"Marta Gogluska-Obirek(Erasmus outgoing Students)",
-        name3:"Angelika Cernitori (Eramus incoming students)",
-        tel:"+49 30 450 576 088 (Lutz Steiner, Leitung)",
-        email1:"erasmixpaten@charite.de" ,
-        address:"Reinhardtstr. 58, CCM",
-        talk:"Sprechzeiten nach Terminvereinbarung",
-        href: 'https://www.charite.de/international/fuer_studierende/' 
-    
-   },
-   {
-        title:"Peer Mentoring",
-        paragraph:"Das Peer-Mentoring-Programm richtet sich an Studienanfängerinnen und -anfänger im ersten und zweiten Semester und soll der Orientierung am Studienbeginn dienen. Hierzu finden regelmäßige Treffen zwischen Mentor bzw. Mentorin und Mentees statt.",
-        name1:"Pinkus Tober-Lau",
-        tel:"+49 30 450 577 259",
-        address:"Virchowweg 24, Aufgang B, 2. Etage, R02.002, CCM",
-        talk:"Sprechzeiten nach Terminvereinbarung",
-        href: 'https://mentoring.charite.de/fuer_studienanfaenger/' 
-   },
-   {
-        title:"Student Mentoring",
-        paragraph:"Das Team des Mentoringprogramms unterstützt Studierende der Charité – Universitätsmedizin Berlin, ihr Studium erfolgreich und zielstrebig zu absolvieren. Im Mentoringprogramm werden Studierenden engagierte Ärztinnen und Ärzte als Mentor*innen vermittelt.",
-        name1:"Petra Meinhardt(Leitung)",
-        name2:"Leonie Kühn und Emma Kula (Organisation)",
-        tel:"+49 30 450 576 091",
-        email1:"studentmentoring@charite.de" ,
-        address:"Hannoversche Straße 19, 3. Etage, CCM",
-        talk:"Sprechzeiten nach Terminvereinbarung",
-        href: 'https://studentmentoring.charite.de' 
-   },
-   {
-        title:"Referat für Studienangelegenheiten",
-        paragraph:"Hier erhalten Sie Informationen und Beratung zu allen allgemeinen und organisatorischen Fragen in puncto Studium.",
-        name1:"Petra Meinhardt(Leitung)",
-        tel:"+49 30 450 576 042 ",
-        email1:"petra.meinhardt@charite.de" ,
-        address:"Hannoversche Str. 19, 3.OG",
-        talk:"Mo-Fr 09:00-12:00, Mo-Do 13:00-16:00(Telefonisch)",
-        talk1:"Di, Do, Fr: 9:30-12:30, Di 13:30-16:00(Sprechzeiten)",
-        href: 'https://www.charite.de/studium_lehre/struktur_einrichtungen/referat_fuer_studienangelegenheiten/' 
-   },
-   {
-        title:"Frauen- und Gleichstellungsbeauftragte",
-        paragraph:"Hier werden Sie aktuell über alle wichtigen Aktivitäten, Maßnahmen und Veranstaltungen informiert, die für die Gleichstellung von Frauen und Männern an der Charité unternommen werden. Unsere Handlungsfelder sind dabei sehr breit gefächert und umfassen Themen wie Karriereplanung, Diversity oder Vereinbarkeit von Familie und Beruf",
-        name1:"Dr. Christine Kurmeyer",
-        name2:"Sekretariat: Stefanie Schmidt ",
-        tel:"+49 30 450 577 252",
-        address:"Westring 1, CVK",
-        talk:"Sprechzeiten nach Terminvereinbarung",
-        href: 'https://frauenbeauftragte.charite.de/' 
-   },
-   {
-        title:"Geschäftsstelle für gute wissenschaftliche Praxis",
-        paragraph:"Die Geschäftsstelle Gute wissenschaftliche Praxis berät bei methodischen Fragen, beispielsweise bei Fragen der Datenerhebung, Bildbearbeitung und der statistischen Auswertung, oder vermittelt den Kontakt zu Expertinnen und Experten die weiterhelfen.",
-        name1:"Claudia Mathan",
-        tel:"+49 30 450 576 056",
-        email1:"gwp@charite.de", 
-        address:"",
-        talk:"Sprechzeiten nach Terminvereinbarung",
-        href: 'https://promotion.charite.de/fuer_promovierende/#c30201158' 
-   },
-   {
-    title:"Fachschaft Initiative Medizin(FSI Medizin)",
-    paragraph:"Die Fachschaftsinitiative ist die Interessenvertretung der Medizinstudierenden an der Charité Universitätsmedizin Berlin. Unter dem Dach der FSI arbeiten zahlreiche Arbeitsgruppen. Darüber hinaus ist die FSI in nahezu allen Gremien an der Charité vertreten. ",
-    name1:"Stella Berboth, Theodor Kempe",
-    tel:"+49 30 450 576050",
-    email1:"sprecher@fsi-charite.de" ,
-    address:"Philippstraße 12, “Hexenhaus”, CCM",
-    talk:"Sprechzeiten nach Terminvereinbarung",
-    href: 'http://fsi-charite.de' 
-}
+    ConsultingCard: [
+        {
+            title: t(`Referat für Studienangelegenheiten`),
+            paragraph: t(`Hier erhalten Sie Informationen und Beratung zu allen allgemeinen und organisatorischen Fragen in puncto Studium.`),
+            tel: "+49 30 450 576 042 ",
+            address: "Hannoversche Str. 19, 3.OG",
+            talk: ["Mo-Fr 09:00-12:00, Mo-Do 13:00-16:00(Telefonisch)", "Di, Do, Fr: 9:30-12:30, Di 13:30-16:00(Sprechzeiten)"],
+            href: 'https://www.charite.de/studium_lehre/struktur_einrichtungen/referat_fuer_studienangelegenheiten/'
+        },
+        {
+            title: t(`Fachschaft Initiative Medizin (FSI Medizin)`),
+            paragraph: t(`Die Fachschaftsinitiative ist die Interessenvertretung der Medizinstudierenden an der Charité Universitätsmedizin Berlin. Unter dem Dach der FSI arbeiten zahlreiche Arbeitsgruppen. Darüber hinaus ist die FSI in nahezu allen Gremien an der Charité vertreten.`),
+            tel: "+49 30 450 576050",
+            emails: ["sprecher@fsi-charite.de"],
+            address: "Philippstraße 12, “Hexenhaus”, CCM",
+            talk: [t(`Sprechzeiten nach Terminvereinbarung`)],
+            href: 'http://fsi-charite.de'
+        },
+        {
+            title: t(`MediCoach`),
+            paragraph: t(`MediCoach bietet ein individuelles, kostenloses, lösungsorientiertes Coaching zur Unterstützung bei Stressbelastungen, die sich direkt aus dem Studium oder auch durch private Lebensumstände ergeben. MediCoach möchte dabei unterstützen, die Studienzeit gewinnbringend und bei guter Gesundheit nutzen zu können. Die Beratungen finden nach Bedarf in ein- oder mehrmaligen Einzelgesprächen oder Gruppensitzungen statt.`),
+            tel: " +49 30 450 529189",
+            emails: ["medicoach@charite.de"],
+            talk: [t(`Gespräch nach Terminvereinbarung`)],
+            href: 'https://medicoach.charite.de/fuer_studierende'
+        },
+        {
+            title: "Peer Mentoring",
+            paragraph: t(`Das Peer-Mentoring-Programm richtet sich an Studienanfänger*innen im ersten und zweiten Semester und soll der Orientierung am Studienbeginn dienen. Dazu finden regelmäßige Treffen zwischen Mentor*in und Mentees statt.`),
+            emails: ["peer-mentoring@charite.de"],
+            talk: [t(`Sprechzeiten nach Terminvereinbarung`)],
+            href: 'https://mentoring.charite.de/fuer_studienanfaenger/'
+        },
+        {
+            title: t(`Familienbüro`),
+            paragraph: t(`Familienbüro:Sie studieren an der Charité und betreuen Kinder? Sie suchen einen Kitaplatz? Sie pflegen Angehörige? Sie haben Fragen und benötigen Informationen rund um das Thema Vereinbarkeit von Studium und Familie?`),
+            tel: "+49 030 450 577 257",
+            emails: ["familienbuero@charite.de", "vaeterbeauftragte@charite.de"],
+            address: "Westring 1, 1. Etage, R1.1011 CVK",
+            talk: ["CVK: dienstags von 11:00 bis 13:00 Uhr, Gästehaus R.0.0120", "CCM: mittwochs von 11:00 bis 13:00 Uhr, Hufelandweg 2, EG, Raum 56|57"],
+            href: 'https://familienbuero.charite.de/charite_mit_kind/studieren_mit_familie/kinderbetreuung'
+        },
+        {
+            title: "Charité International Cooperations (ChiC)",
+            paragraph: t(`ChiC bietet Informationen und Unterstützung rund um die Themen Studieren im Ausland bzw. aus dem Ausland und Unterstützung von internationalen Studierenden in Berlin.`),
+            tel: "+49 30 450 576 088 ",
+            emails: ["erasmixpaten@charite.de"],
+            address: ["Reinhardtstr. 58, CCM"],
+            talk: [t(`Sprechzeiten nach Terminvereinbarung`)],
+            href: 'https://www.charite.de/international/fuer_studierende/'
+        },
+        {
+            title: t(`Student Mentoring`),
+            paragraph: t(`Das Team des Mentoringprogramms unterstützt Studierende, ihr Studium erfolgreich und zielstrebig zu absolvieren. Im Mentoringprogramm werden Studierenden engagierte Ärztinnen und Ärzte als Mentor*innen vermittelt.`),
+            tel: "+49 30 450 576 091",
+            emails: ["studentmentoring@charite.de"],
+            talk: [t(`Sprechzeiten nach Terminvereinbarung`)],
+            href: 'https://studentmentoring.charite.de'
+        },
+        {
+            title: t(`Frauen- und Gleichstellungsbeauftragte`),
+            paragraph: t(`Hier finden Sie Informationen zu allen wichtigen Aktivitäten und Veranstaltungen,die für die Gleichstellung von Frauen und Männern an der Charité unternommen werden. Unsere Handlungsfelder umfassen Themen wie Karriereplanung, Diversity oder Vereinbarkeit von Familie und Beruf.`),
+            tel: "+49 30 450 577 252",
+            address: "Westring 1, CVK",
+            talk: [t(`Sprechzeiten nach Terminvereinbarung`)],
+            href: 'https://frauenbeauftragte.charite.de/'
+        },
+        {
+            title: t(`Schwerbehinderten Vertretung`),
+            paragraph: t(`Die Gesamtschwerbehindertenvertretung (GSBV) der Charité – Universitätsmedizin Berlin vertritt die Interessen der schwerbehinderten oder gleichgestellten Kolleg*innen in der Charité.`),
+            tel: "+49 30 450 576 446",
+            address: "Hindenburgdamm 30, 12203 Berlin, CBF , Geländeadresse: Haus II, 3. Obergeschoss, Raum 302",
+            talk: [t(`Sprechzeiten nach Terminvereinbarung`)],
+            href: ' https://gsbv.charite.de'
+        },
+        {
+            title: t(`Geschäftsstelle für gute wissenschaftliche Praxis`),
+            paragraph: t(`Die Geschäftsstelle Gute wissenschaftliche Praxis berät bei methodischen Fragen,bei Fragen der Datenerhebung, Bildbearbeitung und der statistischen Auswertung, oder vermittelt den Kontakt zu Expert*innen, die weiterhelfen.`),
+            tel: "+49 30 450 576 056",
+            emails: ["gwp@charite.de",],
+            talk: [t(`Sprechzeiten nach Terminvereinbarung`)],
+            href: 'https://promotion.charite.de/fuer_promovierende/#c30201158'
+        }
     ]
-}
- 
+})
